requirements.txt
peewee==3.16.2
pydantic==2.3.0
pydantic_settings==2.0.3
python-dotenv==1.0.0
python-telegram-bot==20.5
python_telegram_bot_calendar==1.0.5
Requests==2.31.0
